+++ webpack/assets/javascripts/react_app/components/common/forms/Submit.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import Button from '../../common/forms/Button';
+
+export default ({ onSubmit, onCancel, disabled = false }) => {
+  return (
+    <div className="clearfix">
+      <div className="form-actions">
+        <Button className="btn-default" disabled={disabled} onSubmit={onCancel}>
+          {__('Cancel')}
+        </Button>
+
+        <Button className="btn-primary" disabled={disabled} onSubmit={onSubmit}>
+          {__('Submit')}
+        </Button>
+
+      </div>
+    </div>
+  );
+};
