--- webpack/assets/javascripts/react_app/components/personalAccessTokens/tokenForm/index.js
@@ -1,18 +1,0 @@
-import React, { Component } from 'react'
-import Button from '../../common/forms/Button';
-
-class TokenForm extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    return (
-      <div>
-      <Button>Submit</Button>
-      </div>
-    )
-  }
-}
-
-export default TokenForm;
