+++ webpack/assets/javascripts/react_app/redux/actions/users/personalAccessTokens/index.js
@@ -1,9 +1,8 @@
-import { SHOW_FORM } from '../../consts';
+import { SHOW_FORM } from '../../../consts';
 
 export const showForm = personalAccessToken => {
   return {
     type: SHOW_FORM,
-    payload: {
-    }
+    payload: {}
   };
-}
+};
