+++ webpack/assets/javascripts/react_app/redux/reducers/users/index.js
@@ -1,0 +1,6 @@
+import personalAccessTokens from './personalAccessTokens';
+import { combineReducers } from 'redux';
+
+export default combineReducers({
+  personalAccessTokens
+});
