+++ webpack/assets/javascripts/react_app/redux/reducers/users/personalAccessTokens/index.js
@@ -1,4 +1,4 @@
-import { SHOW_FORM } from '../../consts';
+import { SHOW_FORM } from '../../../consts';
 import Immutable from 'seamless-immutable';
 
 const initialState = Immutable({
@@ -9,14 +9,6 @@ export default (state = initialState, action) => {
   const { payload } = action;
 
   switch (action.type) {
-    //case TOASTS_ADD: {
-    //  return state.setIn(['messages', payload.key], payload.message);
-    //}
-
-    //case TOASTS_DELETE: {
-    //  return state.set('messages', state.messages.without(payload.key));
-    //}
-
     case SHOW_FORM: {
       return state.set('action', 'form');
     }
